Improve the timetable page for a single train
Flesh out the API and create a timetable for the single-train page
(`timetable/train/:trainId`), displaying basic info about the train as
well as the timetable.

The individual trips (schedules) associated with each train are still
displayed (depending on their presence within the GTFS data), but the
tables are functional.

Still need to work on styling.